worker: unify resume state files and split delivery out of resume.rs
resume.json and finished.json were two files read by two branches,
and the finish transition wrote one then deleted the other. A crash
between the two left both behind, letting a replacement worker
adopt and supervise a build whose result was already packed. One
state.json with a tagged Running/Finished enum makes the transition
a single atomic rename.

The delivery half (registry, persistence, ack, reaper) moves to
resume/delivery.rs so adoption and delivery each read on their own.
